third_party/WebKit/LayoutTests/accessibility/menu-list-popup-reuses-objects-expected.txt

Alicia
Peter
Kalinda

This test makes sure menu lists fire events on both the option that lost selection and the one that gained selection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axMenuListPopup.role is "AXRole: AXMenuListPopup"
PASS axMenuListOption.role is "AXRole: AXMenuListOption"
PASS element.isEqual(axMenuListOption) is true
PASS element.parentElement().isEqual(axMenuListPopup) is true
PASS successfullyParsed is true

TEST COMPLETE
